Set `nvcc` compiler version
To simplify aligning around CUDA version, simply set the
`nvcc_compiler_version` to match the intended CUDA version. This will
ensure that it gets carried over to the recipe during the build. Also as
the compiler version will be correctly set, we will get the expected
`run_export`, which attaches `cudatoolkit` as a dependency with the
right version constraint. This should clean up the recipes a bit.