manually test `py-xgboost` pkg after it's build
For some reason, the `conda build` command currently fails very early on without the `--no-test` flag with an error message about not being able to find the locally built version of `py-xgboost`.

This error message seems to appear before the `py-xgboost` package is actually built, leading me to think that it's a bug with `conda`.

For whatever reason, using the `--no-test` flag seems to get us past that error message.

To account for the `--no-test` flag being added, I added a step to manually test the `py-xgboost` package before it's uploaded.