Fix build scripts
This PR updates the build scripts to change the directory where the built `conda` packages are stored.

Using the `/tmp` directory is not a good idea since that directory is shared between the host machine and container by the `remote-docker-plugin`. This causes problems because artifacts from previous builds may still exist on the host machine in that directory.

Additionally, this PR adds some logging to the build scripts.