update build variables
in Jenkins, we've always built against `rapidsai/xgboost` branches. this commit updates the `XGBOOST_GIT_{REPO,REF}` variables accordingly.

additionally, it updates the `XGBOOST_VERSION` variable to match what we use in Jenkins.

finally, it bumps the build string to `1` so that when this PR is merged, a new package will be uploaded to Anaconda.org