Bump build number
After #51 was merged, only the `3.10` packages were uploaded to Anaconda.org since `3.8`/`3.9` packages with the same build number/version already exist on Anaconda.org

This results in a mix of `3.8`/`3.9` packages that were built with `11.5` and `3.10` packages that were built with `11.8`.

To ensure that a fresh set of packages are uploaded which are all built with `11.8`, this PR bumps the build number for the `xgboost` packages.