Remove duplicate `librmm` runtime dependency
Similarly to https://github.com/rapidsai/raft/pull/1264, this PR removes a duplicate `run` dependency on `librmm`.

The current packages on Anaconda.org have a dependency on the `librmm` packages with the old date string format which have been removed from the `main` label.

Additionally it increments the build number so that new packages are published as a result of this change.